requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.120.0
vaderSentiment==3.3.1
google-api-python-client==2.120.0